models: add video model
