+++ fuocore/models/lyric.py
@@ -1,0 +1,13 @@
+from .base import BaseModel, ModelType
+
+
+class LyricModel(BaseModel):
+    """Lyric Model
+
+    :param SongModel song: song which lyric belongs to
+    :param str content: lyric content
+    :param str trans_content: translated lyric content
+    """
+    class Meta:
+        model_type = ModelType.lyric.value
+        fields = ['song', 'content', 'trans_content']
