+++ fuocore/models/uri.py
@@ -42,6 +42,7 @@ TYPE_NS_MAP = {
     ModelType.playlist: 'playlists',
     ModelType.user: 'users',
     ModelType.lyric: 'lyrics',
+    ModelType.video: 'videos',
 }
 NS_TYPE_MAP = {
     value: key
