+++ fuocore/serializers/python.py
@@ -2,7 +2,8 @@ from .base import Serializer, SerializerMeta, SerializerError, \
     SimpleSerializerMixin
 from .model_helpers import ModelSerializerMixin, SongSerializerMixin, \
     ArtistSerializerMixin, AlbumSerializerMixin, PlaylistSerializerMixin, \
-    UserSerializerMixin, SearchSerializerMixin, ProviderSerializerMixin
+    UserSerializerMixin, SearchSerializerMixin, ProviderSerializerMixin, \
+    VideoSerializerMixin
 
 
 class PythonSerializer(Serializer):
@@ -79,6 +80,11 @@ class ListSerializer(PythonSerializer, metaclass=SerializerMeta):
 
 class SongSerializer(ModelSerializer, SongSerializerMixin,
                      metaclass=SerializerMeta):
+    pass
+
+
+class VideoSerializer(ModelSerializer, VideoSerializerMixin,
+                      metaclass=SerializerMeta):
     pass
